Computing two 32-bit hashes is worse than computing one 64-bit hash.
The reason is that the collision probabilities for each hash are too
high, since the denominator is only 2^32 but the numerator is still
the length of the string. Combining two epsilon [almost] universal
hashes creates a collision probability of at least epsilon^2, so if
epsilon doesn't have a 1 in the numerator, the universality gets
worse.