Add UMASH to the variable-length benchmark
UMASH is a hash function that is also based on the pseudo-dot-product
like CLHASH, VHASH, etc. It has tweaks that were needed for
Backtrace's use cases, including running on input that has a length,
in bytes, that is not a multiple of sizeof(uint64_t) and being very
fast on short input.

Adding benchmarks that demonstrate that speed compared to CLHASH is
future work.

This has been tested with the hash
74ee6984f1f34ea6e5ccc871a3c3c396acf5491c in the UMASH repo.

For more on UMASH, see:

https://engineering.backtrace.io/2020-08-24-umash-fast-enough-almost-universal-fingerprinting